refactor: Move custom streams to submodules of stream
