+++ src/stream/buffered.rs
