+++ tests/async.rs
@@ -28,7 +28,7 @@ use combine::{any, count_min_max, skip_many, Parser, many1};
 use combine::combinator::{any_partial_state, no_partial, optional, recognize, AnyPartialState,
                           skip_many1};
 use combine::stream::RangeStream;
-use combine::easy;
+use combine::stream::easy;
 use combine::parser::char::{char, digit, letter};
 
 quick_error! {
