+++ tests/buffered_stream.rs
@@ -1,11 +1,11 @@
 #![cfg(feature = "std")]
 extern crate combine;
 use combine::stream::IteratorStream;
-use combine::buffered_stream::BufferedStream;
-use combine::easy::Error;
+use combine::stream::buffered::BufferedStream;
+use combine::stream::easy::Error;
 use combine::parser::char::{char, digit, spaces, string};
 use combine::{choice, many, sep_by, try, Parser, Positioned, many1};
-use combine::state::State;
+use combine::stream::state::State;
 
 #[test]
 fn shared_stream_buffer() {
